Refactor to be more easily extensible and readable
