+++ src/extensions/advance_time_extension.ts
@@ -1,0 +1,11 @@
+'use strict';
+
+import { WebviewExtension } from './webview_extension';
+
+export class AdvanceTimeExtension implements WebviewExtension {
+    public generateContent(): string {
+        return `\
+deltaTime = clock.getDelta();
+time = startingTime + clock.getElapsedTime() - pausedTime;`;
+    }
+}
