+++ src/extensions/advance_time_if_not_paused_extension.ts
@@ -1,0 +1,19 @@
+'use strict';
+
+import { WebviewExtension } from './webview_extension';
+
+export class AdvanceTimeIfNotPausedExtension implements WebviewExtension {
+    public generateContent(): string {
+        return `\
+if (paused == false) {
+    deltaTime = clock.getDelta();
+    time = startingTime + clock.getElapsedTime() - pausedTime;
+    vscode.postMessage({
+        command: 'updateTime',
+        time: time
+    });
+} else {
+    deltaTime = 0.0;
+}`;
+    }
+}
