+++ src/extensions/audio/audio_pause_extension.ts
@@ -1,0 +1,9 @@
+'use strict';
+
+import { WebviewExtension } from '../webview_extension';
+
+export class AudioPauseExtension implements WebviewExtension {
+    public generateContent(): string {
+        return `audioContext.suspend();`;
+    }
+}
