+++ src/extensions/audio/audio_resume_extension.ts
@@ -1,0 +1,9 @@
+'use strict';
+
+import { WebviewExtension } from '../webview_extension';
+
+export class AudioResumeExtension implements WebviewExtension {
+    public generateContent(): string {
+        return `audioContext.resume();`;
+    }
+}
