+++ src/extensions/audio/no_audio_extension.ts
@@ -1,0 +1,12 @@
+'use strict';
+
+import { WebviewExtension } from '../webview_extension';
+
+export class NoAudioExtension implements WebviewExtension {
+    public generateContent(): string {
+        return `\
+const audioContext = {
+    sampleRate: 0
+};`;
+    }
+}
