+++ src/extensions/pause_whole_render_extension.ts
@@ -1,0 +1,9 @@
+'use strict';
+
+import { WebviewExtension } from './webview_extension';
+
+export class PauseWholeRenderExtension implements WebviewExtension {
+    public generateContent(): string {
+        return `if (paused) return;`;
+    }
+}
