+++ src/extensions/textures/texture_extension_extension.ts
@@ -1,0 +1,5 @@
+'use strict';
+
+export interface TextureExtensionExtension {
+    generateTextureContent(): string;
+}
