+++ src/extensions/user_interface/screenshot_button_extension.ts
@@ -1,0 +1,9 @@
+'use strict';
+
+import { WebviewExtension } from '../webview_extension';
+
+export class ScreenshotButtonExtension implements WebviewExtension {
+    public generateContent(): string {
+        return `<span id="screenshot"></span>`;
+    }
+}
