+++ src/extensions/webview_extension.ts
@@ -1,0 +1,5 @@
+'use strict';
+
+export interface WebviewExtension {
+    generateContent(): string;
+}
